requirements.txt
accelerate==0.27.2
asttokens
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
comm
contourpy==1.2.0
cycler==0.12.1
debugpy
decorator
diffusers==0.26.3
einops==0.7.0
exceptiongroup
executing
filelock==3.13.1
fonttools==4.48.1
fsspec==2024.2.0
gdown==5.1.0
huggingface-hub==0.21.3
idna==3.6
importlib-metadata
ipykernel
ipython
jedi
Jinja2==3.1.3
jupyter_client
jupyter_core
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline
mpmath==1.3.0
nest_asyncio
networkx==3.2.1
numpy==1.26.4
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.19.3
# nvidia-nvjitlink-cu12==12.3.101
# nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging
parso
pexpect
pickleshare
pillow==10.2.0
platformdirs
prompt-toolkit
psutil
ptyprocess
pure-eval
Pygments
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil
PyYAML==6.0.1
pyzmq
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
setuptools==69.0.3
six
soupsieve==2.5
stack-data
sympy==1.12
tokenizers==0.15.2
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado
tqdm==4.66.2
traitlets
transformers==4.38.2
#triton==2.2.0
typing_extensions
urllib3==2.2.0
wcwidth
wheel==0.42.0
zipp
